pyproject.toml
[project]
name = "enterprise-ai-platform"
version = "0.1.0"
description = "Enterprise AI Adoption Platform with RBAC, RAG, and multi-tenant support"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Jason", email = "jason@matherly.net"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    # Database
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Redis
    "redis>=5.2.0",
    # Vector DB
    "qdrant-client>=1.13.0",
    # Numerical computing (for semantic cache similarity)
    "numpy>=1.26.0",
    # Azure AI
    "azure-ai-projects>=1.0.0b5",
    "azure-identity>=1.19.0",
    "openai>=1.59.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    # Observability
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-client>=0.21.0",
    "langfuse>=3.12.0",
    "structlog>=24.4.0",
    # Document Processing
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "tiktoken>=0.8.0",
    # Utilities
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "tenacity>=9.0.0",
    "greenlet>=3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

# ============================================
# Build System (Hatchling)
# ============================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Tell hatch where to find the source code (src layout)
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/pyproject.toml",
    "/README.md",
]

# ============================================
# uv Configuration
# ============================================
[tool.uv]
# Use precompiled Python from python-build-standalone
python-preference = "managed"

# ============================================
# Dependency Groups (replaces tool.uv.dev-dependencies)
# ============================================
[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "psycopg[binary]>=3.2.0",
]

# ============================================
# Ruff Configuration (Linting & Formatting)
# ============================================
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T10",  # flake8-debugger
    "ISC",  # flake8-implicit-str-concat
    "PIE",  # flake8-pie
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "ARG",  # flake8-unused-arguments
    "ERA",  # eradicate (commented code)
    "PL",   # pylint
    "PERF", # perflint
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "S101",    # assert in tests is fine
    "PLR0913", # too many arguments
    "PLW0603", # global statement (intentional singleton pattern)
    "PLC0415", # import outside top-level (intentional lazy loading/circular import avoidance)
    "PLR2004", # magic values (low priority, can address later)
    "PLR0915", # too many statements (refactoring scope)
    "PLR0912", # too many branches (refactoring scope)
    "ARG001",  # unused function argument (sometimes intentional for API consistency)
    "ARG002",  # unused method argument (sometimes intentional)
    "S110",    # try-except-pass (sometimes intentional silent error handling)
    "PERF401", # manual list comprehension (low priority optimization)
    "PERF403", # manual dict comprehension (low priority optimization)
    "SIM102",  # collapsible if (readability preference)
    "ERA001",  # commented out code (sometimes useful for documentation)
    "B008",    # function call in default argument (FastAPI pattern with Depends/File)
    "N818",    # exception naming (style preference)
    "S104",    # binding to all interfaces (intentional for dev/container networking)
    "S105",    # hardcoded password variable names (false positive - loaded from env)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================
# MyPy Configuration (Type Checking)
# ============================================
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/", "spikes/"]

# ============================================
# Pytest Configuration
# ============================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
